Revert "Moves paged kv cache kernel to kernels dir (#2131)"
This reverts commit 51a223e7674b1c1e3f3067465ec703ba83c278ab.
