+++ sharktank/sharktank/kernels/__init__.py
@@ -17,4 +17,3 @@ from .pooling_nchw_sum import *
 from .base import *
 from .bitcast import *
 from .topk import *
-from .kv_cache import *
